Many changes to media section, almost complete.

+++ Media/views.py
@@ -13,13 +13,16 @@ from Connexion.models import stagiare
 
 def cours(request):
    if request.user.username :
+        profil_type = request.session['profil_type']
         cours_list = Cours.objects.all()
-        context = {'cours_list': cours_list}
+        context = {'cours_list': cours_list,
+                   'profil_type':profil_type,
+                   }
         return render(request,'Media/cours.html',context)
    else :
         return redirect("/accueil/login/")
 
-def cours_page(request):
+def cours_page(request,cours_id):
      if request.user.username :
          cours = Cours.objects.get(id=cours_id)
          context = {'cours':cours}
@@ -29,8 +32,9 @@ def cours_page(request):
 
 def cas_cliniques(request):
    if request.user.username :
+         profil_type = request.session['profil_type']
          cas_cliniques_list = CasCliniques.objects.all()
-         context = {'cours_list': cas_cliniques_list}
+         context = {'cours_list': cas_cliniques_list, 'profil_type':profil_type,}
          return render(request, 'Media/cas_cliniques.html',context)
    else : 
         return redirect("/accueil/login/")
@@ -48,7 +52,9 @@ class CoursCreateView(CreateView):
     model= Cours
     fields = ['title','description','category','support_pdf','video_src']
    
-def get_success_url(self):
+    def get_success_url(self):
+            self.object.author=self.request.user.username
+            self.object.save()
             return reverse('cours-page', kwargs={'cours_id' : self.object.pk})
         
 
@@ -57,14 +63,41 @@ class CasCliniquesCreateView(CreateView):
     fields = ['title','description','category','support_pdf','corrige_src','video_src']
 
     def get_success_url(self):
+        self.object.author=self.request.user.username
+        self.object.save()
         return reverse('cas-clinique-page', kwargs={'cas_clinique_id' : self.object.pk})
 
 
+def addCours(request):
+    if request.user.username :
+        profil_type = request.session['profil_type']
+        if(profil_type=="formateur"):
+            return CoursCreateView.as_view()(request)
+        else:
+            return HttpResponse("Cette opération est Impossible, ACCES DENIED!")
+
+def addCasCliniques(request):
+    if request.user.username :
+        profil_type = request.session['profil_type']
+        if(profil_type=="formateur"):
+            return CasCliniquesCreateView.as_view()(request)
+        else:
+            return HttpResponse("Cette opération est Impossible, ACCES DENIED!")
 
 def removeCours(request,cours_id):
-    Cours.objects.filter(id=cours_id).delete()
-    return cours(request)
+    if request.user.username :
+        profil_type = request.session['profil_type']
+        if(profil_type=="formateur"):
+            Cours.objects.filter(id=cours_id).delete()
+            return cours(request)
+        else:
+            return HttpResponse("Cette opération est Impossible, ACCES DENIED!")
 
 def removeCasCliniques(request,cours_id):
-    CasCliniques.objects.filter(id=cours_id).delete()
-    return cas_cliniques(request)
+    if request.user.username :
+        profil_type = request.session['profil_type']
+        if(profil_type=="formateur"):
+            CasCliniques.objects.filter(id=cours_id).delete()
+            return cours(request)
+        else:
+            return HttpResponse("Cette opération est Impossible, ACCES DENIED!")
